refactor(test): extract shared cascade test helpers to conftest
Move _FakePeer, _FakeDiscovery, _FakeSpawnAdapter, _make_drive_loop,
and _make_agent_task from test_cascade.py and test_cascade_delegation.py
into tests/test_ravn/conftest.py. Both test files now import the shared
helpers, eliminating verbatim duplication.

Also refactor TestDriveLoopHeartbeat to use _make_drive_loop() instead
of manually constructing InitiativeConfig + Settings + DriveLoop.
